Initial step towards bulk edit
